model/requirements.txt
numpy 
pandas
scikit-learn
matplotlib
ipykernel
seaborn 
streamlit
bayesian-optimization
optuna
plotly.express
spc_spectra
openpyxl
flask
lightgbm
hyperopt
xgboost
